requirements.txt
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.1.0
playwright==1.49.0
python-telegram-bot==21.7
pyyaml==6.0.1
apscheduler==4.0.0a5
python-dotenv==1.0.0
